Added local signing of ethereum transactions made by the KM node.
This commit is huge, as it is a squash of several week's worth of work
done by myself (@reuvenpo) and @cashmaney. we do this because some of
the commits done by cashmaney were not signed/verified and the branch
configuration refused merging it all.

In this work we change the way ethereum functions are made against the
Enigma smart contract from the KM node, such that we no longer rely on
an unlocked account on the ethereum node we connect to, but rather sign
each request with a private key stored in the KM, which is associated
with a new ethereu address.

There are two possible ways of generating the keys and address for the KM node:

1. We generate a new keypair and address inside the SGX enclave, and expose the
   public key
2. We read the private key and address from the configuration.

One of these is necessarily chosen.